Added program exit values for gpu-engine and gpu-memclock. This is mainly relevant for mining rigs that are also used for other purposes. Like my gaming PC. :)
When my R9 290 is mining, it is running at 1000 engine and 1500 memclock. But when I am using it for gaming, it runs 1117 engine and 1375 memclock. Normally I have to use a utility like MSI AfterBurner or Sapphire TriXX to do so. But this change allows me to set an exit value for each GPU when cgminer exits.

Above functionality as an example: --gpu-engine 1000:1117 --gpu-memclock 1500:1375

You can still combine it with multiple cards like this:
--gpu-engine 900,1000:1117 --gpu-memclock 1375,1500:1375
or this:
--gpu-engine 900:880,1000:1117 --gpu-memclock 1375:1250,1500:1375